+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Zombienet SDK integration tests for elastic scaling on Asset Hub Polkadot and People Polkadot.
 - collectives-polkadot-runtime: Increase the Secretary Collective salary budget ([#1172](https://github.com/polkadot-fellows/runtimes/pull/1172))
 
 ## [2.2.2] 23.04.2026